change queryIndexes to unsigned int all around. Observe each queryIndex without respect to if it aligned or not.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@31934 60cd2ccc-2b05-0410-9566-da273a940bff

+++ cpp/src/goby/C_Reads.cc
@@ -205,7 +205,7 @@ extern "C" {
      * Do NOT free the char *'s you send to this method. If you need to keep a copy, MAKE A COPY.
      * @return the Goby read index
      */
-    unsigned long gobyReads_nextSequence(
+    unsigned int gobyReads_nextSequence(
         CReadsHelper *readsHelper,
         char **readIdentifierpp, char **descriptionpp,
         char **sequencepp, int *sequenceLength,
@@ -257,7 +257,7 @@ extern "C" {
      * Do NOT free the char *'s you send to this method. If you need to keep a copy, MAKE A COPY.
      * @return the Goby read index
      */
-    unsigned long gobyReads_nextSequencePair(
+    unsigned int gobyReads_nextSequencePair(
         CReadsHelper *readsHelper,
         char **readIdentifierpp, char **descriptionpp,
         char **sequencepp, int *sequenceLength,
